+++ CHANGELOG.md
@@ -1,1 +1,1 @@
-- added an option to display direction arrows (vanilla feature), to show if the player is above or below wherever you're looking
+- fixed force waypoints option